draws empty grid, green
